% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{call_function}
\alias{call_function}
\title{Call a SQL function/procedure.}
\usage{
call_function(con, schema, function_name, args, dialect = "T-SQL", cast = TRUE)
}
\arguments{
\item{con}{A database connection.}

\item{schema}{A string, the schema to query.}

\item{function_name}{A string, the function/procedure to query.}

\item{args}{A named list or vector, names are the parameter names and values are the parameter values.}

\item{dialect}{A string, "T-SQL" or "Postgresql".,}

\item{cast}{TRUE/FALSE, if TRUE, will add SQL to cast the parameters to the specified type.}
}
\value{
A data.table.
}
\description{
Call a SQL function/procedure.
}
\examples{
call_function(con = NULL)
}
