% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{connect}
\alias{connect}
\title{Connect to a database using a connection string via DBI/odbc.}
\usage{
connect(
  con_str = "Driver={PostgreSQL ANSI};Host=localhost;Port=5432;Database=postgres;"
)
}
\arguments{
\item{con_str}{A database connection string.}
}
\value{
A database connection.
}
\description{
Connect to a database using a connection string via DBI/odbc.
}
\examples{
connect(NULL)
}
