% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{drop_table}
\alias{drop_table}
\title{Generate a DROP TABLE statement, optionally execute the statement if con is not NULL.}
\usage{
drop_table(args, con = NULL)
}
\arguments{
\item{args}{A string, the arguments to add to the DROP TABLE statement.}

\item{con}{A database connection that can be passed to DBI::dbSendQuery/DBI::dbGetQuery.}
}
\value{
A string, the DROP TABLE statement; or the results retrieved by DBI::dbSendQuery after executing the statement.
}
\description{
Generate a DROP TABLE statement, optionally execute the statement if con is not NULL.
}
\examples{
drop_table("sample")
}
