% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{fetch_columns}
\alias{fetch_columns}
\title{Retrieve the columns/types in a table.}
\usage{
fetch_columns(con, schema, table)
}
\arguments{
\item{con}{A database connection.}

\item{schema}{A string, the schema to query.}

\item{table}{A string, the table to query.}
}
\value{
A data.table.
}
\description{
Retrieve the columns/types in a table.
}
\examples{
fetch_columns(con = NULL)
}
