% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{fetch_tables}
\alias{fetch_tables}
\title{Retrieve the tables in a schema}
\usage{
fetch_tables(con, schema)
}
\arguments{
\item{con}{A database connection.}

\item{schema}{A string, the schema to query.}
}
\value{
A data.table.
}
\description{
Retrieve the tables in a schema
}
\examples{
fetch_tables(con = NULL)
}
