% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDbFromFile.R
\name{queryDbFromFile}
\alias{queryDbFromFile}
\title{Run a query against a database}
\usage{
queryDbFromFile(server, database, filename, ...)
}
\arguments{
\item{server}{The name of the DSN for the server you want to connect to}

\item{database}{The name of the database that you want to run the query against}

\item{filename}{A path to a file to read the query from.}

\item{...}{Named parameters to interpolate into the query}
}
\value{
The results of the query as a tibble
}
\description{
This function allows you to run a query against a database, returning the
results in a tibble.
}
\details{
The query is read from the file that is passed as an argument. See
\link{queryDb} if you want to pass the query as a string instead.

The connection to the server is made by odbc and uses the name of a DSN to
connect to. This removes any requirement to hard code in usernames, passwords
and connection strings into files.

Parameters can be used in queries by either specifying each parameter and
it's value after the query parameter, or by specifying all of the parameters
in a named list.
}
\examples{
\dontrun{
queryDbFromFile("MyServerName", "MyDatabase", "file.sql")

queryDbFromFile("MyServerName",
                "MyDatabase",
                "file.sql",
                id = 1,
                date = Sys.Date())

params <- list(id = 1, date = Sys.Date())
queryDbFromFile("MyServerName",
                "MyDatabase",
                "file.sql",
                params)
}
}
