% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{sqlizeTypes}
\alias{sqlizeTypes}
\title{Get the equivalent SQL data type for a given R object.}
\usage{
sqlizeTypes(x, dialect = "T-SQL")
}
\arguments{
\item{x}{A R object.}

\item{dialect}{A string, "T-SQL" or "Postgresql".}
}
\value{
A string, the equivalent SQL data type for x.
}
\description{
Get the equivalent SQL data type for a given R object.
}
\examples{
sqlizeTypes(100.1209)
}
