% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{t_sql_script_create_table}
\alias{t_sql_script_create_table}
\title{Generate a CREATE TABLE statement for an existing table in Microsoft SQL Server.}
\usage{
t_sql_script_create_table(con, table)
}
\arguments{
\item{con}{A database connection that can be passed to DBI::dbSendQuery/DBI::dbGetQuery.}

\item{table}{A string, the schema qualified table name of an existing SQL table.}
}
\value{
A data table, contains the DDL scripts for creating a table.
}
\description{
Generate a CREATE TABLE statement for an existing table in Microsoft SQL Server.
}
\examples{
t_sql_script_create_table(con = NULL)
}
