% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{truncate_table}
\alias{truncate_table}
\title{Generate a TRUNCATE TABLE statement, optionally execute the statement if con is not NULL.}
\usage{
truncate_table(args, con = NULL)
}
\arguments{
\item{args}{A string, the arguments to add to the TRUNCATE TABLE statement.}

\item{con}{A database connection that can be passed to DBI::dbSendQuery/DBI::dbGetQuery.}
}
\value{
A string, the TRUNCATE TABLE statement; or the results retrieved by DBI::dbGetQuery after executing the statement.
}
\description{
Generate a TRUNCATE TABLE statement, optionally execute the statement if con is not NULL.
}
\examples{
truncate_table(args = "table1")
}
